Deeply refactor Makefile
These changes make all targets parameterized around the experiment
"strategy" and "version". This allows these elements to be listed
once at the tope of the file to generate all relevant build rules.

This abstraction is not pursued for its own sake, but rather to enable
the generation of targets of the form

    experiment-<strategy>-<version>

Where a "strategy" is currently either `001indinv` or `002bmc`.

A target of this form will run all the experiments for given strategy
and version.

These targets will be used in the CI pipeline in a build matrix
parameterized on the version and strategy, which should allow us to run
each experiment suite on a separate build server.